Cleaned-up examples

+++ python/src/mdedevicecontrol/dcgraph.py
@@ -234,6 +234,9 @@ class Graph:
 
     async def delete_device_configuration(self,policy_id):
 
+        if policy_id is None:
+            logger.error("policy_id is None")
+            return RuntimeError()
         logger.debug("policy id="+policy_id)
         try:
             result = await self.graph_client.device_management.device_configurations.by_device_configuration_id(policy_id).delete()
